+++ CHANGELOG.md
@@ -35,8 +35,10 @@ and this project adheres to
 ### Fixed
 
 - Fix `CurveData` values alignment in the buffer (#216).
-- Fix `CurveData` buffer initialization (#217).
+- Fix `CurveData` buffer initialization (#217, #218).
 - Fix `CClass::ToString` address hash (#216).
+- Fix `TTypedClass::Assign` to act as copy assignment instead of copy construction (#218).
+- Fix `RawBuffer` move construction and move assignment (#218).
 
 ## [1.0.0] - 2026-03-09